requirements.txt
boto3
certifi==2024.7.4
cockroachdb==0.3.5
flask
Flask==3.0.3
flask-apispec==0.11.4
Flask-Cors==5.0.0
Flask-Testing==0.8.1
hubspot-api-client
marshmallow==3.21.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
prisma
psycopg2-binary==2.9.10
pytz==2024.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.31
sqlalchemy-cockroachdb==2.0.2
stripe==9.1.0
Werkzeug==3.0.6